gis/pyproject.toml
[project]
name = "gis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [

    #GEE
    "earthengine-api>=1.0.0",
    "geopandas>=1.1.1",
    "shapely>=2.1.2",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.5",
    
    #Environmental loading
    "python-dotenv>=1.0.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]